advanced 리팩토링
